requirements.txt
streamlit>=1.36
python-docx>=1.1.0
openai>=1.40.0